Change infrastructure for SCTP_MAX_BURST to allow compliance with the latest socket API ID. Especially it can be disabled.
Full compliance needs changing the structure used in the
socket option. Since this breaks the API, it will be a
seperate commit which will not be MFCed to stable/8.

MFC after: 3 months.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@217894 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/netinet/sctp_pcb.h
@@ -317,7 +317,7 @@ struct sctp_pcb {
 	uint32_t initial_sequence_debug;
 	uint32_t adaptation_layer_indicator;
 	uint32_t store_at;
-	uint8_t max_burst;
+	uint32_t max_burst;
 	char current_secret_number;
 	char last_secret_number;
 };
